scraper/requirements.txt
# Core scraping
scrapy==2.11.0
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
aiohttp==3.9.1
httpx==0.25.2

# Database
pymongo==4.6.1
motor==3.3.2
supabase==2.0.0

# Task Scheduling
apscheduler==3.10.4

# Data Processing
pydantic==2.5.0
python-dateutil==2.8.2
pandas==2.1.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
fake-useragent==1.4.0
tenacity==8.2.3
cloudscraper==1.2.71

# Image Processing
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.12.0
flake8==6.1.0
